requirements.txt
google-generativeai>=0.3.0
pydantic>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
supabase-py==2.3.0 